src/behavioral/mediator/ChatRoom.java
package behavioral.mediator;

import java.util.ArrayList;
import java.util.List;

/**
 * Concrete Mediator implementing the ChatMediator interface
 */
public class ChatRoom implements ChatMediator {
    private List<User> users;

    public ChatRoom() {
        this.users = new ArrayList<>();
    }

    @Override
    public void sendMessage(String message, User user) {
        for (User u : users) {
            // Message should not be received by the user sending it
            if (u != user) {
                u.receive(message);
            }
        }
    }

    @Override
    public void addUser(User user) {
        users.add(user);
    }
}
